Crate mpot

Re-exports

pub use mailpot::melib::smol;

Modules

anyhow

github crates-io docs-rs

args 🔒

changesets

Changeset structs: update specific struct fields.

log

A lightweight logging facade.

mail

Types for processing new posts: PostFilter, ListContext, MailJob and PostAction.

melib

A crate that performs mail client operations such as

message_filters

Filters to pass each mailing list post through. Filters are functions that implement the PostFilter trait that can:

models

Database models: MailingList, ListOwner, ListSubscription, PostPolicy, SubscriptionPolicy and Post.

postfix

Generate configuration for the postfix mail server.

rusqlite

Rusqlite is an ergonomic wrapper for using SQLite from Rust.

serde

Serde

serde_json

Serde JSON

submission

Submit e-mail through SMTP.

Macros

list 🔒

string_opts 🔒

Structs

Account

An account entry.

AccountChangeset

Changeset struct for Account.

Configuration

The configuration for the mailpot database and the mail server.

Connection

A connection to a mailpot database.

DbVal

A database entry and its primary key. Derefs to its inner type.

Envelope

Envelope represents all the header and structure data of an email we need to know.

EnvelopeHash

Error

The Error type.

ListCandidateSubscription

A mailing list subscription candidate.

ListContext

List context passed to a list’s PostFilter stack.

ListOwner

A mailing list owner entry.

ListOwnerChangeset

Changeset struct for ListOwner.

ListSubscription

A mailing list subscription entry.

ListSubscriptionChangeset

Changeset struct for ListSubscription.

MailingList

A mailing list entry.

MailingListChangeset

Changeset struct for Mailinglist.

MailtoAddress

A mailto: value.

PostPolicy

A mailing list post policy entry.

QueueEntry

A queue entry.

SubscriptionPolicy

A mailing list subscription policy entry.

Template

A named template.

TemplateRenderContext

Helper type for Connection::send_reply_with_list_template.

Enums

ErrorKind

The kind of an error.

ListRequest

Type of mailing list request.

MailJob

Scheduled jobs added to a ListContext by a list’s PostFilter stack.

PostAction

Post action returned from a list’s PostFilter stack.

Queue

In-database queues of mail.

SendMail

How to send e-mail.

Traits

ChainedError

This trait is implemented on all the errors generated by the error_chain macro.

Context

Provides the context method for Result.

MaildirPathTrait

ResultExt

Additional methods for Result, for easy interaction with this crate.

Functions

get_file_hash 🔒

main 🔒

run_app 🔒

submit 🔒

Type Definitions

Result

Convenient wrapper around std::Result.